pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "snn-research"
version = "17.1.0"
description = "A Next-Generation Neuromorphic Computing framework simulating Artificial Brains with Spiking Neural Networks, Transformers, and Cognitive Architectures."
readme = "README.md"
requires-python = ">=3.10"
keywords = [
    "snn",
    "neuromorphic-computing",
    "spiking-neural-networks",
    "artificial-brain",
    "cognitive-architecture",
    "pytorch",
    "spikformer",
    "artificial-intelligence",
    "neuroscience"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
dependencies = [
    "click",
    "torch",
    "torchvision",
    "torchaudio",
    "gradio==4.20.0",
    "flask",
    "numpy",
    "pandas",
    "scikit-learn",
    "matplotlib",
    "seaborn",
    "pillow",
    "snntorch",
    "spikingjelly",
    "langchain",
    "langchain-community",
    "openai",
    "transformers",
    "sentence-transformers",
    "datasets",
    "omegaconf",
    "hydra-core",
    "dependency-injector",
    "tqdm",
    "jupyter",
    "ipywidgets",
    "requests",
    "beautifulsoup4",
    "typer==0.16.0",
    "PyYAML",
    "h5py",
    "GPyOpt",
    "GPy",
    "cupy==13.6.0",
    "networkx",
    "pandas-stubs",
    "tqdm",
    "plotly",
    "sentence-transformers",
]

[project.urls]
Homepage = "https://github.com/matsushibadenki/snn"
Documentation = "https://github.com/matsushibadenki/snn/blob/main/README.md"
Repository = "https://github.com/matsushibadenki/snn.git"

[project.scripts]
snn-cli = "snn_research.cli.main:main"

[tool.setuptools.packages.find]
include = ["snn_research*", "app*"]

[tool.mypy]
show_error_context = true
show_column_numbers = true
pretty = true